Control.Par.Class

Description

This module establishes a class hierarchy that captures the interfaces of Par monads. There are two layers: simple futures (ParFuture) and full IVars (ParIVar). All Par monads are expected to implement the former, some also implement the latter.

For more documentation of the programming model, see

The Control.Monad.Par module in the monad-par package.

The wiki/tutorial (http://www.haskell.org/haskellwiki/Par_Monad:_A_Parallelism_Tutorial)

The original paper (http://www.cs.indiana.edu/~rrnewton/papers/haskell2011_monad-par.pdf)

Tutorial slides (http://community.haskell.org/~simonmar/slides/CUFP.pdf)

Other slides (http://www.cs.ox.ac.uk/ralf.hinze/WG2.8/28/slides/simon.pdf, http://www.cs.indiana.edu/~rrnewton/talks/2011_HaskellSymposium_ParMonad.pdf)

The essence of Par monads: forking control flow

class (Functor p, Monad p) => ParMonad p where

The essence of a Par monad is that its control flow is a binary tree of forked threads.

Note, this class also serves a secondary purpose similar: providing an implementation-internal way to lift IO into tho Par monad. However, this is a different use case than either MonadIO or ParThreadSafe. Unlike the latter, ALL Par monads should be a member of this class. Unlike the former, the user should not be able to access the internalLiftIO operation of this class from Safe code.

Methods

fork :: p () -> p ()

Forks a computation to happen in parallel.

Futures: basic parallelism with communication

class ParMonad m => ParFuture m where

ParFuture captures the class of Par monads which support futures. This level of functionality subsumes par/pseq and is similar to the Control.Parallel.Strategies.Eval monad.

A minimal implementation consists of spawn_ and get. However, for monads that are also a member of ParIVar it is typical to simply define spawn in terms of fork, new, and put.

Associated Types

type Future m :: * -> *

The type of a future that goes along with the particular Par monad the user chooses.

type FutContents m a :: Constraint

Different implementations may place different constraints on what is allowable inside a Future. For example, some implementations require an Eq Constraint.

Methods

spawn :: (NFData a, FutContents m a) => m a -> m (Future m a)

Create a potentially-parallel computation, and return a future (or promise) that can be used to query the result of the forked computataion.

  spawn p = do
    r <- new
    fork (p >>= put r)
    return r

spawn_ :: FutContents m a => m a -> m (Future m a)

Like spawn, but the result is only head-strict, not fully-strict.

get :: FutContents m a => Future m a -> m a

Wait for the result of a future, and then return it.

spawnP :: (NFData a, FutContents m a) => a -> m (Future m a)

Spawn a pure (rather than monadic) computation. Fully-strict.

  spawnP = spawn . return

IVars: futures that anyone can fill

class ParFuture m => ParIVar m where

ParIVar builds on futures by adding full anyone-writes, anyone-reads IVars. These are more expressive but may not be supported by all distributed schedulers.

A minimal implementation consists of fork, put_, and new.

Methods

new :: forall a. FutContents m a => m (Future m a)

creates a new IVar new :: m (IVar m a)

put :: forall a. (FutContents m a, NFData a) => IVar m a -> a -> m ()

put a value into a IVar. Multiple puts to the same IVar are not allowed, and result in a runtime error.

put fully evaluates its argument, which therefore must be an instance of NFData. The idea is that this forces the work to happen when we expect it, rather than being passed to the consumer of the IVar and performed later, which often results in less parallelism than expected.

Sometimes partial strictness is more appropriate: see put_.

put_ :: forall a. FutContents m a => IVar m a -> a -> m ()

like put, but only head-strict rather than fully-strict.

newFull :: (FutContents m a, NFData a) => a -> m (IVar m a)

creates a new IVar that contains a value

newFull_ :: FutContents m a => a -> m (IVar m a)

creates a new IVar that contains a value (head-strict only)

Data structures that can be consumed in parallel

class Generator c => ParFoldable c where

Collection types which can be consumed in parallel.

Methods

pmapFold

Arguments

:: forall m a t . (ParFuture m, FutContents m a)

=> (ElemOf c -> m a)
compute one result

-> (a -> a -> m a)
combine two results

-> a
initial accumulator value

-> c
element generator to consume

-> m a

class Generator c where

Collections that can generate a sequence of elements of the same element type.

Reason: We have a problem where some types (like Ranges) are splittable, but they are not containers for arbitrary data. Thus we introduce a more limited concept of a data source that can generate only a particular kind of element (but cannot be constructed or traversed).

It is trivial to provide an instance for any type that is already a Functor:

 import Data.Foldable as F
 instance Foldable f => Generator (f a) where
   type ElemOf (f a) = a
   foldrM = F.foldrM 

However, we don't provide this blanket instance because it would conflict with more tailored instances that may be desired for particular containers. For example, a Data.Map generator might include keys as well as values.

Finally, note that a much more general version of this class can be found in Data.Generator from the reducers package.

Associated Types

type ElemOf c :: *

Methods

fold :: (acc -> ElemOf c -> acc) -> acc -> c -> acc

Fold all outputs from the generator, sequentially. The ordering is determined by the generator type, and is whatever the natural and inexpensive ordering for that type is.

In general this should be used with commutative, associative functions. This fold is strict in the accumulator.

foldM :: Monad m => (acc -> ElemOf c -> m acc) -> acc -> c -> m acc

A monadic version of fold.

foldMP :: ParMonad m => (acc -> ElemOf c -> m acc) -> acc -> c -> m acc

If the monad in question is a Par monad, then this version can sometimes be more efficient than foldM.

forM_ :: Monad m => c -> (ElemOf c -> m ()) -> m ()

Execute an action for each output of the generator.

forMP_ :: ParMonad m => c -> (ElemOf c -> m ()) -> m ()

The same as forM_ but for a Par monad.

(Internal) Abstracting LVar Schedulers.

class (Monad m, ParSealed m) => LVarSched m where

Abstract over LVar-capable schedulers. This is not for end-user programming.

Associated Types

type LVar m :: * -> * -> *

The type of raw LVars, shared among all specific data structures (Map, Set, etc).

Methods

newLV :: IO a -> m (LVar m a d)

Create an LVar, including the extra state that required for scheduling and synchronization.

putLV

Arguments

:: LVar m a d
the LVar

-> (a -> IO (Maybe d))
how to do the put, and whether the LVar's value changed

-> m ()

Update an LVar. The change itself happens as an IO action, but any state changes must be linearizable and must respect the lattice semantics for the LVar.

getLV

Arguments

:: LVar m a d
the LVar

-> (a -> Bool -> IO (Maybe b))
already past threshold? The Bool indicates whether the LVar is FROZEN.

-> (d -> IO (Maybe b))
does d pass the threshold?

-> m b

Do a threshold read on an LVar. This requires both a global and delta handler, and if the underlying Par monad assumes idempotency, both could even execute for the same value.

stateLV :: LVar m a d -> (Proxy (m ()), a)

Extract a handle on the raw, mutable state within an LVar.

forkLV :: m () -> m ()

Fork a child thread.

yield :: m ()

Put ourselves at the bottom of the work-pile for the current thread, allowing others a chance to run.

schedCheck :: m ()

Usually a no-op. This checks in with the scheduler, so that it might perform any tasks which need to be performed periodically.

returnToSched :: m a

Drop the current continuation and return to the scheduler.

class (Monad qm, LVarSched m, ParQuasi m qm) => LVarSchedQ m qm | m -> qm where

An LVar scheduler with the quasi-determinism capability. This interface is complicated by the fact that there are two monads (deterministic and quasideterministic).

Methods

freezeLV :: LVar m a d -> (Proxy (m ()), qm ())

Freeze an LVar (introducing quasi-determinism). This requires marking it at runtime.

class (Monad m, Functor m, Monad qm, Functor qm) => ParQuasi m qm | m -> qm where

Par monads which can be switched into a quasi-deterministic variant. (See Kuper et al. POPL 2014).

Methods

toQPar :: m a -> qm a

Lift a deterministic computation to a quasi-deterministic one.

class ParSealed m

All proper Par monads should have an s parameter that seals them so that live Futures, IVars, etc cannot escape from a Par computation. Membership in this class provides a uniform way to extract these s parameters where needed.

Associated Types

type GetSession m :: *

Extract the s parameter from a Par monad type.

data Proxy a

Carry a phantom type argument.

Constructors

Proxy

Simple tracking of WHICH Par monads permit only threadsafe effects

class Monad p => ParThreadSafe p

The class of Par monads in which all monadic actions are threadsafe and do not care which thread they execute on. Thus it is ok to inject additional parallelism.

class NFData a

A class of types that can be fully evaluated.