Control.Par.Class.Unsafe

Description

Unsafe operations that end users should NOT import.

This is here only for other trusted implementation components.

Documentation

class Monad p => ParThreadSafe p where

The class of Par monads in which all monadic actions are threadsafe and do not care which thread they execute on. Thus it is ok to inject additional parallelism.

Methods

unsafeParIO :: IO a -> p a

Run some IO in parallel on whatever thread we happen to be on. The end user does not get access to this.

class (Functor p, Monad p) => ParMonad p where

The essence of a Par monad is that its control flow is a binary tree of forked threads.

Note, this class also serves a secondary purpose similar: providing an implementation-internal way to lift IO into tho Par monad. However, this is a different use case than either MonadIO or ParThreadSafe. Unlike the latter, ALL Par monads should be a member of this class. Unlike the former, the user should not be able to access the internalLiftIO operation of this class from Safe code.

Methods

fork :: p () -> p ()

Forks a computation to happen in parallel.

internalLiftIO :: IO a -> p a

(Internal! Not exposed to the end user.) Lift an IO operation. This should only be used by other infrastructure-level components, e.g. the implementation of monad transformers or LVars.