Control.Par.EffectSigs

Description

Type-level effect signatures that characterize the side-effects of a given Par computation.

A type-level record of effects

data EffectSig

Currently-tracked effects. This is INTERNAL/private datatype and subject to change. You should use constraints such as HasGet rather than directly depending on this, whereever possible.

Constructors

Ef Putting Getting Freezing Bumping IOing

Simple boolean flags, each with a disttinct kind to prevent mixups

data Putting

A boolean by another name.

Constructors

P

NP

data Getting

A boolean by another name.

Constructors

G

NG

data Freezing

A boolean by another name.

Constructors

F

NF

data Bumping

A boolean by another name.

Constructors

B

NB

data IOing

A boolean by another name.

Constructors

I

NI

User-visible Constraints

type HasPut e = GetP e ~ P

type HasGet e = GetG e ~ G

type HasFreeze e = GetF e ~ F

type HasBump e = GetB e ~ B

type HasIO e = GetI e ~ I

type NoPut e = NP ~ GetP e

type NoGet e = NG ~ GetG e

type NoFreeze e = NF ~ GetF e

type NoBump e = NB ~ GetB e

type NoIO e = NI ~ GetI e

Derived constraints, i.e. shorthands for common combinations:

type ReadOnly e = (NoPut e, NoBump e, NoFreeze e, NoIO e)

type Deterministic e = (NoFreeze e, NoIO e)

type QuasiDeterministic e = NoIO e

type Idempotent e = (NoBump e, NoIO e)

Manipulating the phantom types of a Par monad

type family GetEffects m :: EffectSig

Type-level utility function for extracting the e part of a valid Par-monad stack.

type family SetEffects e m :: * -> *

Type-level utility function for replacing the s part of a valid Par-monad stack.

type family GetSession m :: *

Type-level utility function for extracting the s part of a valid Par-monad stack.

type family SetSession s m :: * -> *

Type-level utility function for replacing the s part of a valid Par-monad stack.

Accessor and setter functions for EffectSigs

type family GetP e :: Putting

Utility for getting just the Put effect flag.

type family GetG e :: Getting

Utility for getting just the Get effect flag.

type family GetF e :: Freezing

Utility for getting just the Freeze effect flag.

type family GetB e :: Bumping

Utility for getting just the Bump effect flag.

type family GetI e :: IOing

Utility for getting just the IO effect flag.

type family SetP p e :: EffectSig

Utility for setting just the Put effect flag.

type family SetG p e :: EffectSig

Utility for setting just the Get effect flag.

type family SetF p e :: EffectSig

Utility for setting just the Freeze effect flag.

type family SetB b e :: EffectSig

Utility for setting just the Bump effect flag.

type family SetI b e :: EffectSig

Utility for setting just the IO effect flag.

Lifted constraints and type functions directly on monads

type family SetMP p m :: * -> *

Shorthand for setting the Put effeect of a Par monad.

type family SetMG g m :: * -> *

Shorthand for setting the Get effeect of a Par monad.

type family SetMF f m :: * -> *

Shorthand for setting the Freeze effeect of a Par monad.

type family SetMB b m :: * -> *

Shorthand for setting the Bump effeect of a Par monad.

type family SetMI i m :: * -> *

Shorthand for setting the IO effeect of a Par monad.

type HasIOM m = HasIO (GetEffects m)

type ReadOnlyM m = ReadOnly (GetEffects m)

type IdempotentM m = Idempotent (GetEffects m)