Data.Splittable.Class

Description

A simple type class for data that can be split into pieces for parallel operation, and then reassembled.

Documentation

class Eq a => Split a where

Data that can be split into balanced pieces. The main application of this is parallel consumption of the data.

Methods

split :: a -> [a]

Split the data value into pieces. An empty data structure may return an empty list.

splitPlease :: Int -> a -> [a]

A variant of split that allows the user to provide a hint as to how many pieces they would like to split into. There is no obligation for the implementation to follow this hint (either as an upper or lower bound).